Move databaseAccess out of the routes folder, as it has nothing to do with routing
